agent/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic>=0.49.0
duckduckgo-search==7.3.2
python-dotenv==1.0.1
kernel>=0.34.0
apscheduler>=3.10.0
Pillow>=10.0.0
httpx>=0.27.0
